% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmgarbage.R
\name{rmg_bad_alnum_ratio}
\alias{rmg_bad_alnum_ratio}
\title{Rule A: alphanumeric ratio}
\usage{
rmg_bad_alnum_ratio(x, threshold = 0.5)
}
\arguments{
\item{x}{a character vector}

\item{threshold}{a value between 0 and 1 for the alphanumeric to puncutation ratio}
}
\value{
Logical
}
\description{
If a string's ratio of alphanumeric characters to total
characters is less than 50\%, the string is garbage.
From https://github.com/foodoh/rmgarbage
}
\examples{

garbage_string_r1 <-
  paste0(stringi::stri_rand_strings(1, 2, pattern = "[a-zA-Z0-9]"),
        "14.9˜;tv˜;<F~~~~9er&é(er@@|^˜:..", # this is garbage
          collapse = "")

garbage_string_r2 <-
  stringi::stri_rand_strings(1, 10, pattern = "[a-zA-Z0-9]")

  rmg_bad_alnum_ratio(garbage_string_r1) # TRUE
  rmg_bad_alnum_ratio(garbage_string_r2) # FALSE

}
