% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmgarbage.R
\name{rmg_bad_consonant_vowel_ratio}
\alias{rmg_bad_consonant_vowel_ratio}
\title{Rule V: consonants to vowels ratio}
\usage{
rmg_bad_consonant_vowel_ratio(x, ratio = 0.1)
}
\arguments{
\item{x}{a character vector}

\item{ratio}{numeric, ratio of vowels to consonents}
}
\value{
Logical


}
\description{
If a string has nothing but alphabetic characters, look at the
number of consonants and vowels. If the number of one is less than 10\%
of the number of the other, then the string is garbage.
From https://github.com/foodoh/rmgarbage
}
\examples{

 garbage_string_cv1 <- "azerytugino"
 garbage_string_cv2 <- "aaaaaaaaaaaaaaaaaax"

 rmg_bad_consonant_vowel_ratio(garbage_string_cv1) # FALSE
 rmg_bad_consonant_vowel_ratio(garbage_string_cv2) # TRUE

}
