% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmgarbage.R
\name{rmg_consecutive_four_identical}
\alias{rmg_consecutive_four_identical}
\title{Rule R: four identical characters}
\usage{
rmg_consecutive_four_identical(x, n = 4L)
}
\arguments{
\item{x}{a character vector}

\item{n}{integer, how identical consequtive characters to detect}
}
\value{
Logical
}
\description{
If a string has 4 identical characters in a row, it is garbage.
From https://github.com/foodoh/rmgarbage
}
\examples{

garbage_string_4c1 <-  "aaaaazzzezrertrg"
garbage_string_4c2 <-  "azertyuiopqsdfghj"

rmg_consecutive_four_identical(garbage_string_4c1) # TRUE
rmg_consecutive_four_identical(garbage_string_4c2) # FALSE


}
