% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmgarbage.R
\name{rmg_has_two_distinct_puncts_inside}
\alias{rmg_has_two_distinct_puncts_inside}
\title{Rule P: punctuation characters in a string}
\usage{
rmg_has_two_distinct_puncts_inside(x, last_n = 2)
}
\arguments{
\item{x}{a character vector}

\item{last_n}{numeric, number of final characters to strip}
}
\value{
Logical
}
\description{
Strip off the first and last characters of a string. If there
are two distinct punctuation characters in the result, then the string
is garbage. Stripping off the last two characters as false positives
includes those ending with ').' and similar.
}
\examples{

garbage_string_2p1 <- "'hi!cou?cou,"
garbage_string_2p2 <-  "jean-marc,"

rmg_has_two_distinct_puncts_inside(garbage_string_2p1) # TRUE
rmg_has_two_distinct_puncts_inside(garbage_string_2p2) # FALSE

}
