% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmgarbage.R
\name{rmg_has_uppercase_within_lowercase}
\alias{rmg_has_uppercase_within_lowercase}
\title{Rule C: internal capitals}
\usage{
rmg_has_uppercase_within_lowercase(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
Logical
}
\description{
If a string begins and ends with a lowercase letter, then if
the string contains an uppercase letter anywhere in between,
then it is removed as garbage.
}
\examples{

garbage_string_u1 <- "stratigraphic"
garbage_string_u2 <- "sUatigraphic"

rmg_has_uppercase_within_lowercase(garbage_string_u1) # FALSE
rmg_has_uppercase_within_lowercase(garbage_string_u2) # TRUE


}
