% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmgarbage.R
\name{rmg_toolong}
\alias{rmg_toolong}
\title{Rule L: Too long}
\usage{
rmg_toolong(x, length = 20)
}
\arguments{
\item{x}{a character vector}

\item{length}{numeric, threshold for acceptable length of string}
}
\value{
Logical
}
\description{
Tests whether the string passed is more than 20 characters,
if yes. It is garbage! From https://github.com/foodoh/rmgarbage
}
\examples{
garbage_string_n39 <- stringi::stri_rand_strings(1, 39, pattern = "[a-z]")
garbage_string_n40 <- stringi::stri_rand_strings(1, 40, pattern = "[a-z]")

rmg_toolong(garbage_string_n39)  # FALSE
rmg_toolong(garbage_string_n40)  # TRUE

}
