% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmgarbage.R
\name{rmg_uppercase_lowercase_ratio}
\alias{rmg_uppercase_lowercase_ratio}
\title{Upper to lowercase ratio}
\usage{
rmg_uppercase_lowercase_ratio(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
Logical
}
\description{
If the number of uppercase characters
in a string is greater than the number of lowercase
characters, and if the number of uppercase characters
is less than the total number of characters in the string,
it is garbage. From https://ladyissy.github.io/OCR/
}
\examples{

garbage_string_ul1 <- "UNESCO"
garbage_string_ul2 <- "POTAto"

rmg_uppercase_lowercase_ratio(garbage_string_ul1) # FALSE
rmg_uppercase_lowercase_ratio(garbage_string_ul2) # TRUE


}
