% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmgarbage.R
\name{rmgarbage}
\alias{rmgarbage}
\title{Apply all tests}
\usage{
rmgarbage(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
Logical
}
\description{
Applys all the tests to detect any kind of OCR garbage
}
\examples{
rmgarbage("Tptpmn") # TRUE
rmgarbage("UNESCO") # FALSE
rmgarbage("bmarwick@uw.edu") # FALSE


}
